Initial Commit
This is a bare minimum HYP firmware to boot into the aarch32
LK (little kernel) bootloader in EL1. For now the return address
is hardcoded to 0x8f600000 (entry point of LK), because it's not
clear where to get the entry point from, yet!

(Note: I have written most of the commit messages *after* the
 firmware was already done to document its creation process...)

The code I've written is mostly based on the "ARM Architecture
Reference Manual for Armv8-A" and similar EL2/EL1 initialization
sequences adapted from Linux and U-Boot.

Surprisingly flashing this dead simple firmware to the "hyp"
partition actually gets us to LK. With this, it is easy to dump
the initial registers to memory, light up GPIO LEDSs to see when
something runs or not and so on. :D