Bring RPM out of reset on initial boot-up
Looking closer at the Linux kernel log I notice that all RPM regulators
seem to be suddenly broken:

  adv7511 3-0039: failed to init regulators

Poking some registers it looks like RPM is still in reset, because
0x1860000 (GCC_APSS_MISC) still has the reset bit set. Apparently,
the HYP firmware is also responsible for bringing RPM out of reset
on the initial boot-up.

Implementing this isn't that hard, so add some additional assembly
to store if this is the initial boot and clear the reset bit
if necessary.

This makes the RPM regulators work again! :)