Cleanup HCR_EL2/VBAR_EL2 before jumping to aarch64 bootloader in EL2
At the moment, attempting to boot U-Boot in EL2 using the HVC call
results in the device rebooting. I'm not entirely sure why that happens,
but it stops happening if I clear the HCR_EL2 register before jumping
to U-Boot. I guess U-Boot does not properly reset these registers as
it should. (Because the ARM manual describes most registers as having
"an architecturally UNKNOWN" value on reset...)

So, reset HCR_EL2 and while we're at it also VBAR_EL2 before we jump
to the aarch64 bootloader/kernel in EL2.