Use SMC call even for booting initial bootloader in aarch64 EL2
So far I have always tested with LK (little kernel) as primary bootloader,
which runs in aarch32 state. The HYP firmware should also be able to jump
to a aarch64 bootloader (like U-Boot) directly in EL2. However, for some
reason the U-Boot port for the DragonBoard 410c was designed to be booted
from LK instead of using it as primary bootloader in EL1.

Actually, it's quite easy to get U-Boot booting as primary bootloader.
This works partially even with Qualcomm's original HYP firmware:

 1. Change the following two config options in U-Boot:
     CONFIG_SYS_TEXT_BASE=0x8F600000
     CONFIG_OF_EMBED=y
 2. Strip the generated U-Boot ELF binary ($OUT/u-boot)
 3. Sign it like the LK bootloader
 4. Flash it to the "aboot" partition

It boots directly the the U-Boot command prompt. Nice!

Unfortunately, with Qualcomm's original HYP firmware - and even with
qhypstub in its current state - using U-Boot as primary bootloader
results in the same "PSCI bug" that causes all secondary CPU cores
to be started in aarch32 state unless we invoke the SMC call for the
execution state switch in TZ once. This means that again, the board
reboots as soon as we get to "smp: Bringing up secondary CPUs ..."
in Linux. Oh well.

Actually, we have most of the code already to workaround this issue.
For the HVC call that switches from aarch32 to aarch64 EL2 we invoke
the SMC call and force execution back into EL2 to avoid the bug.

The difference in case of a primary aarch64 bootloader is that there
won't be a HVC call. Instead, we need to set up the SMC call directly
from the HYP firmware. So - only on initial boot - avoid jumping
directly to an aarch64 bootloader in EL2 and instead setup and invoke
the SMC call . Then, reuse the stage 2 address translation "hack"
to force execution back to EL2. Works great!