Ensure all or no CPUs are running aarch64 in EL2
At the moment, Linux still complains about "CPUs started in
inconsistent modes" if someone forgets to patch LK and therefore uses
the SMC call instead of the HVC call to switch to aarch64 state.

It doesn't seem to cause any real issues but it doesn't look very nice.
Actually, since the switch to aarch64 EL2 needs to happen through our
HVC call, we can record there that it completed successfully and store
that in a variable. (The existing "execution_state" works just fine
for this...)

Then, whenever we are in the HYP entry point we can check the variable
and only jump to aarch64 directly in EL2 if:

  - The HVC call was invoked at some point, or
  - The primary bootloader runs directly in aarch64 EL2

If this is not the case, i.e. the primary bootloader is aarch32 and
the HVC call was never invoked, then we also jump to aarch64 in EL1.
In that case, all CPUs will be consistently started in EL1.