Initialize SCTLR_EL1 for aarch32 "guests"
At the moment, booting the 3.10 downstream kernel from Qualcomm works
fine with qhypstub with arm64, but crashes early with armv7. Granted,
there is very little reason to still use 32-bit downstream kernel on
the DragonBoard 410c.

But there are also other MSM8916 devices, like the BQ Aquaris X5
(Longcheer L8910) smartphone that can run qhypstub. The BQ Aquaris X5
also runs mainline really well, but occasionally the original vendor
kernel is still useful. And it happens to run on Linux 3.10 in 32-bit.

So, for easier debugging, booting 32-bit Qualcomm Linux on DB410c also
results in a crash very early during boot. Fortunately, it still triggers
the LK crash handler so it's fairly easy to investigate:

  booting linux @ 0x80008000, ramdisk @ 0x84000000 (1364491), tags/device tree @ 0x83e00000
  undefined abort, halting
  r0  0x00000000 r1  0x80100c0e r2  0x00000800 r3  0x80004000
  r4  0x80008000 r5  0x0014d20b r6  0x0000000e r7  0x00000000
  r8  0x83e00000 r9  0x80000000 r10 0x83200000 r11 0x10101105
  r12 0x800080b0 usp 0x00000000 ulr 0x00000000 pc  0x800083d0
  spsr 0x400001d3

The registers are not particularly interesting, except for the program
counter (pc) that tells us on which instruction it crashed. We need to
look at the instruction with offset 0x800083d0 - 0x80008000 = 0x3d0.

Looking at the disassembly of the vmlinuz binary this is part of the
decompression code. It fails when executing the following instruction:

  3d0:       ee070f9a        mcr     15, 0, r0, cr7, cr10, {4}

This is a (for aarch32) very typical access to CP15 registers
(not sure how they are called exactly). They are really badly readable
compared to equivalent instructions in aarch64, but at least there is
"Table G7-3" in the ARM Architecture Reference Manual that explains
what all those magic numbers mean.

Looking through the table, this is a CP15DSB instruction (c7, c10, 4).
Clicking the reference to the explanation of the instruction there is
a big warning that this instruction is deprecated, and its use is
strongly discouraged. Very suspicious.

Looking for more occurences of "CP15DSB" in the document the problem
becomes obvious quite quickly. There is the SCTLR_EL1.CP15BEN register
bit that controls if the deprecated memory barrier instructions are
enabled for EL1 and EL0. And indeed, dumping the register from LK
it seems to be set when using Qualcomm's firmware but not with qhypstub.

The SCTLR_EL1 register can be quite confusing because some bits have
different meanings depending if EL1 runs in aarch64 state or aarch32
state. For aarch64, the initial value of the register is defined to be
"architecturally UNKNOWN", so the arm64 Linux initializes that register
properly. But on aarch32, there are defined reset values, and CP15BEN
should be set to 1 by default. I guess this is why 32-bit ARM Linux
just assumes that this bit is already set.

Therefore, for aarch32 "guests", initialize the SCTLR_EL1 register
from qhypstub and make sure that CP15BEN is set. This makes 32-bit
ARM Linux boot correctly.