Initialize VTCR_EL2/VTTBR_el2 for temporary stage 2 address translation
Apparently, I was really "lucky" to get the stage 2 address translation
"hack" working in the previous commit. I was relying on a special
(consistent but randomly looking) reset value of VTTBR_EL2. Attempting
to set VTTBR_EL2 to < 2^40 breaks aarch64 boots and causes reboots.

VTTBR_EL2 points to the stage 2 address translation table. Normally
we would set up a proper page table (with only invalid entries),
but I wanted to figure out why is is working with VTTBR_EL2 >= 2^40.

Staring at the "ARM Architecture Reference Manual for Armv8-A" for a few
hours (especially the VTTBR_EL2/VTCR_EL2 registers and the part on the
stage 2 address translation) I believe I'm relying on a "Address Size
Fault" which happens if the VTTBR_EL2 address is larger than the physical
address size configured in VTCR_EL2.

And indeed, the physical address size in VTCR_EL2 happens to default
to 40-bits. If I change the physical address size to 32-bit, suddenly
VTTBR_EL2 >= 2^32 is working. It does not really matter what we set up
here as long as it causes the Instruction Abort eventually, so I chose:

  - VTCR_EL2 = 32-bit physical/translated address, 4 KB granule, start at level 1
  - VTTBR_EL2 = 2^32

I guess at this point it is kind of overengineered, but I kind of enjoyed
investigating this. I hope no one will complain as long as it works. :D